Avoid reading and encoding image data as base64 when he data is already base64 encoded.
Also use DataContent.Uri in the above case so that any subsequent calls to evaluate the same image can use the Uri value computed in previous calls instead of recomputing each time.
