+++ src/Libraries/Microsoft.Extensions.AI.Evaluation.Safety/ChatResponseExtensions.cs
@@ -1,0 +1,10 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.Extensions.AI.Evaluation.Safety;
+
+internal static class ChatResponseExtensions
+{
+    internal static bool ContainsImage(this ChatResponse response)
+        => response.Messages.ContainsImage();
+}
